Setup git-ps to do stacked reviews on github.
This will allow us to have stacked reviews like in gerrit more easily
and send a chain.

We can use front ends for github like https://reviewstack.dev/ (or many
others, but this is a nice one maintained by the folks at facebook) to
see these in a nice way like gerrit provides.

<!-- ps-id: 701e56ad-7ddc-49c9-9ba2-a6fc464ed872 -->